Virtual keypads. Move code to separate files, outside the ui_driver.
