+++ mchf-eclipse/files.mak
@@ -129,3 +129,4 @@ drivers/audio/freedv_test_data.c \
 drivers/audio/rtty.c \
 drivers/audio/psk.c \
 drivers/ui/lcd/ui_lcd_layouts.c \
+drivers/ui/ui_vkeybrd.c \
